Make camera daemon stable and failsave

+++ php-scripts/camera-daemon.php
@@ -11,15 +11,21 @@ $conn = OpenCon();
 if ($conn->connect_errno) {
     echo "Errno: " . $conn->connect_errno . "\n";
 }
+//cleanup on fatal error
+$callback = function() use($hLock) {
+	if(file_exists(__FILE__.".lock")) {
+		flock($hLock, LOCK_UN);
+		fclose($hLock);
+		unlink(__FILE__.".lock");
+	}
+};
+register_shutdown_function($callback);
 
-while(true)
+while(file_exists(__FILE__.".lock"))
 {
 	saveAllCams();
     //avoid CPU exhaustion, adjust as necessary
     usleep(29000000);//29 seconds
 }
 
-flock($hLock, LOCK_UN);
-fclose($hLock);
-unlink(__FILE__.".lock");
 ?>
